@@ -15,99 +15,99 @@
  */
 package org.febit.rectify.util;
 
-import jakarta.annotation.Nullable;
 import lombok.experimental.UtilityClass;
+import org.jspecify.annotations.Nullable;
 
 @UtilityClass
 public class Args {
 
     @Nullable
-    public static Integer int0(Object[] args) {
+    public static Integer int0(@Nullable Object @Nullable [] args) {
         return asInt(arg0(args));
     }
 
     @Nullable
-    public static Integer int1(Object[] args) {
+    public static Integer int1(@Nullable Object @Nullable [] args) {
         return asInt(arg1(args));
     }
 
     @Nullable
-    public static Integer int2(Object[] args) {
+    public static Integer int2(@Nullable Object @Nullable [] args) {
         return asInt(arg2(args));
     }
 
     @Nullable
-    public static Integer int3(Object[] args) {
+    public static Integer int3(@Nullable Object @Nullable [] args) {
         return asInt(arg3(args));
     }
 
     @Nullable
-    public static String string0(Object[] args) {
+    public static String string0(@Nullable Object @Nullable [] args) {
         return asString(arg0(args));
     }
 
     @Nullable
-    public static String string1(Object[] args) {
+    public static String string1(@Nullable Object @Nullable [] args) {
         return asString(arg1(args));
     }
 
     @Nullable
-    public static String string2(Object[] args) {
+    public static String string2(@Nullable Object @Nullable [] args) {
         return asString(arg2(args));
     }
 
     @Nullable
-    public static String string3(Object[] args) {
+    public static String string3(@Nullable Object @Nullable [] args) {
         return asString(arg3(args));
     }
 
     @Nullable
-    public static Object arg0(Object[] args) {
+    public static Object arg0(@Nullable Object @Nullable [] args) {
         return argX(args, 0, null);
     }
 
     @Nullable
-    public static Object arg0(Object[] args, @Nullable Object def) {
+    public static Object arg0(@Nullable Object @Nullable [] args, @Nullable Object def) {
         return argX(args, 0, def);
     }
 
     @Nullable
-    public static Object arg1(Object[] args) {
+    public static Object arg1(@Nullable Object @Nullable [] args) {
         return argX(args, 1, null);
     }
 
     @Nullable
-    public static Object arg1(Object[] args, @Nullable Object def) {
+    public static Object arg1(@Nullable Object @Nullable [] args, @Nullable Object def) {
         return argX(args, 1, def);
     }
 
     @Nullable
-    public static Object arg2(Object[] args) {
+    public static Object arg2(@Nullable Object @Nullable [] args) {
         return argX(args, 2, null);
     }
 
     @Nullable
-    public static Object arg2(Object[] args, @Nullable Object def) {
+    public static Object arg2(@Nullable Object @Nullable [] args, @Nullable Object def) {
         return argX(args, 2, def);
     }
 
     @Nullable
-    public static Object arg3(Object[] args) {
+    public static Object arg3(@Nullable Object @Nullable [] args) {
         return argX(args, 3, null);
     }
 
     @Nullable
-    public static Object arg3(Object[] args, @Nullable Object def) {
+    public static Object arg3(@Nullable Object @Nullable [] args, @Nullable Object def) {
         return argX(args, 3, def);
     }
 
     @Nullable
-    public static Object argX(Object[] args, int x) {
+    public static Object argX(@Nullable Object @Nullable [] args, int x) {
         return argX(args, x, null);
     }
 
     @Nullable
-    public static Object argX(@Nullable Object[] args, int x, @Nullable Object def) {
+    public static Object argX(@Nullable Object @Nullable [] args, int x, @Nullable Object def) {
         if (args != null && x >= 0 && x < args.length) {
             return args[x];
         }
@@ -127,11 +127,11 @@ public class Args {
         if (value == null) {
             return null;
         }
-        if (value instanceof Integer) {
-            return (Integer) value;
+        if (value instanceof Integer i) {
+            return i;
         }
-        if (value instanceof Number) {
-            return ((Number) value).intValue();
+        if (value instanceof Number number) {
+            return number.intValue();
         }
         if (value instanceof Character) {
             return (int) ((Character) value);
