@@ -20,8 +20,8 @@ import org.febit.lang.func.Function1;
 import org.febit.lang.func.Function2;
 import org.febit.lang.util.ConvertUtils;
 import org.febit.rectify.lib.IFunctions;
+import org.jspecify.annotations.Nullable;
 
-import javax.annotation.Nullable;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -34,7 +34,10 @@ import java.time.temporal.Temporal;
 import java.time.temporal.TemporalAdjuster;
 import java.time.temporal.TemporalAdjusters;
 
-@SuppressWarnings({"unused"})
+@SuppressWarnings({
+        "java:S1118", // Utility classes should not have public constructors
+        "unused",
+})
 public class TemporalFunctions implements IFunctions {
 
     /**
@@ -46,70 +49,93 @@ public class TemporalFunctions implements IFunctions {
     public static class Proto {
 
         public final Function0<Instant> now = Instant::now;
-        public final Function1<Object, ZoneOffset> zone = ConvertUtils::toZone;
+        public final Function1<Object, @Nullable ZoneOffset> zone = ConvertUtils::toZone;
 
-        public final Function2<String, String, Temporal> parse = TemporalFunctions::parse;
-        public final Function2<Temporal, String, String> format = TemporalFunctions::format;
+        public final Function2<@Nullable String, @Nullable String, @Nullable Temporal> parse
+                = TemporalFunctions::parse;
+        public final Function2<@Nullable Temporal, @Nullable String, @Nullable String> format
+                = TemporalFunctions::format;
 
-        public final Function1<Object, Long> toMillis = ConvertUtils::toMillis;
-        public final Function1<Object, Instant> toInstant = ConvertUtils::toInstant;
+        public final Function1<@Nullable Object, @Nullable Long> toMillis
+                = ConvertUtils::toMillis;
+        public final Function1<@Nullable Object, @Nullable Instant> toInstant
+                = ConvertUtils::toInstant;
 
-        public final Function1<Object, Integer> toHour = ConvertUtils::toHour;
-        public final Function1<Object, LocalTime> toTime = ConvertUtils::toTime;
-        public final Function1<Object, LocalDate> toDate = ConvertUtils::toDate;
-        public final Function1<Object, LocalDateTime> toDateTime = ConvertUtils::toDateTime;
-        public final Function1<Object, Integer> toDateNumber = ConvertUtils::toDateNumber;
-        public final Function1<Object, ZonedDateTime> toZonedDateTime = ConvertUtils::toZonedDateTime;
+        public final Function1<@Nullable Object, @Nullable Integer> toHour
+                = ConvertUtils::toHour;
+        public final Function1<@Nullable Object, @Nullable LocalTime> toTime
+                = ConvertUtils::toTime;
+        public final Function1<@Nullable Object, @Nullable LocalDate> toDate
+                = ConvertUtils::toDate;
+        public final Function1<@Nullable Object, @Nullable LocalDateTime> toDateTime
+                = ConvertUtils::toDateTime;
+        public final Function1<@Nullable Object, @Nullable Integer> toDateNumber
+                = ConvertUtils::toDateNumber;
+        public final Function1<@Nullable Object, @Nullable ZonedDateTime> toZonedDateTime
+                = ConvertUtils::toZonedDateTime;
 
-        public final Function1<Object, Integer> toUtcHour = ConvertUtils::toUtcHour;
-        public final Function1<Object, LocalTime> toUtcTime = ConvertUtils::toUtcTime;
-        public final Function1<Object, LocalDate> toUtcDate = ConvertUtils::toUtcDate;
-        public final Function1<Object, LocalDateTime> toUtcDateTime = ConvertUtils::toUtcDateTime;
-        public final Function1<Object, Integer> toUtcDateNumber = ConvertUtils::toUtcDateNumber;
-        public final Function1<Object, ZonedDateTime> toUtcZonedDateTime = ConvertUtils::toUtcZonedDateTime;
+        public final Function1<@Nullable Object, @Nullable Integer> toUtcHour
+                = ConvertUtils::toUtcHour;
+        public final Function1<@Nullable Object, @Nullable LocalTime> toUtcTime
+                = ConvertUtils::toUtcTime;
+        public final Function1<@Nullable Object, @Nullable LocalDate> toUtcDate
+                = ConvertUtils::toUtcDate;
+        public final Function1<@Nullable Object, @Nullable LocalDateTime> toUtcDateTime
+                = ConvertUtils::toUtcDateTime;
+        public final Function1<@Nullable Object, @Nullable Integer> toUtcDateNumber
+                = ConvertUtils::toUtcDateNumber;
+        public final Function1<@Nullable Object, @Nullable ZonedDateTime> toUtcZonedDateTime
+                = ConvertUtils::toUtcZonedDateTime;
 
-        public final Function2<Temporal, Integer, Temporal> addMillis = plusFunc(ChronoUnit.MILLIS);
-        public final Function2<Temporal, Integer, Temporal> addSeconds = plusFunc(ChronoUnit.SECONDS);
-        public final Function2<Temporal, Integer, Temporal> addMinutes = plusFunc(ChronoUnit.MINUTES);
-        public final Function2<Temporal, Integer, Temporal> addHours = plusFunc(ChronoUnit.HOURS);
-        public final Function2<Temporal, Integer, Temporal> addDays = plusFunc(ChronoUnit.DAYS);
-        public final Function2<Temporal, Integer, Temporal> addMonths = plusFunc(ChronoUnit.MONTHS);
-        public final Function2<Temporal, Integer, Temporal> addYear = plusFunc(ChronoUnit.YEARS);
+        public final Function2<@Nullable Temporal, @Nullable Integer, @Nullable Temporal> addMillis
+                = plusFunc(ChronoUnit.MILLIS);
+        public final Function2<@Nullable Temporal, @Nullable Integer, @Nullable Temporal> addSeconds
+                = plusFunc(ChronoUnit.SECONDS);
+        public final Function2<@Nullable Temporal, @Nullable Integer, @Nullable Temporal> addMinutes
+                = plusFunc(ChronoUnit.MINUTES);
+        public final Function2<@Nullable Temporal, @Nullable Integer, @Nullable Temporal> addHours
+                = plusFunc(ChronoUnit.HOURS);
+        public final Function2<@Nullable Temporal, @Nullable Integer, @Nullable Temporal> addDays
+                = plusFunc(ChronoUnit.DAYS);
+        public final Function2<@Nullable Temporal, @Nullable Integer, @Nullable Temporal> addMonths
+                = plusFunc(ChronoUnit.MONTHS);
+        public final Function2<@Nullable Temporal, @Nullable Integer, @Nullable Temporal> addYear
+                = plusFunc(ChronoUnit.YEARS);
 
-        public final Function1<Temporal, Temporal> firstDayOfMonth
+        public final Function1<@Nullable Temporal, @Nullable Temporal> firstDayOfMonth
                 = adjustFunc(TemporalAdjusters.firstDayOfMonth());
-        public final Function1<Temporal, Temporal> lastDayOfMonth
+        public final Function1<@Nullable Temporal, @Nullable Temporal> lastDayOfMonth
                 = adjustFunc(TemporalAdjusters.lastDayOfMonth());
-        public final Function1<Temporal, Temporal> firstDayOfNextMonth
+        public final Function1<@Nullable Temporal, @Nullable Temporal> firstDayOfNextMonth
                 = adjustFunc(TemporalAdjusters.firstDayOfNextMonth());
 
-        public final Function1<Temporal, Temporal> firstDayOfYear
+        public final Function1<@Nullable Temporal, @Nullable Temporal> firstDayOfYear
                 = adjustFunc(TemporalAdjusters.firstDayOfYear());
-        public final Function1<Temporal, Temporal> lastDayOfYear
+        public final Function1<@Nullable Temporal, @Nullable Temporal> lastDayOfYear
                 = adjustFunc(TemporalAdjusters.lastDayOfYear());
-        public final Function1<Temporal, Temporal> firstDayOfNextYear
+        public final Function1<@Nullable Temporal, @Nullable Temporal> firstDayOfNextYear
                 = adjustFunc(TemporalAdjusters.firstDayOfNextYear());
-    }
 
-    private static Function1<Temporal, Temporal> adjustFunc(TemporalAdjuster adjuster) {
-        return temporal -> {
-            if (temporal == null) {
-                return null;
-            }
-            return temporal.with(adjuster);
-        };
-    }
+        private static Function1<@Nullable Temporal, @Nullable Temporal> adjustFunc(TemporalAdjuster adjuster) {
+            return temporal -> {
+                if (temporal == null) {
+                    return null;
+                }
+                return temporal.with(adjuster);
+            };
+        }
 
-    private static Function2<Temporal, Integer, Temporal> plusFunc(ChronoUnit unit) {
-        return (temporal, i) -> {
-            if (temporal == null) {
-                return null;
-            }
-            if (i == null) {
-                return temporal;
-            }
-            return temporal.plus(i, unit);
-        };
+        private static Function2<@Nullable Temporal, @Nullable Integer, @Nullable Temporal> plusFunc(ChronoUnit unit) {
+            return (temporal, i) -> {
+                if (temporal == null) {
+                    return null;
+                }
+                if (i == null) {
+                    return temporal;
+                }
+                return temporal.plus(i, unit);
+            };
+        }
     }
 
     private static DateTimeFormatter fmt(String pattern) {
