@@ -16,21 +16,19 @@
 package org.febit.rectify.flink;
 
 import org.apache.flink.api.common.typeinfo.BasicTypeInfo;
-import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
-import org.apache.flink.types.Row;
 import org.febit.rectify.format.JsonSourceFormat;
 import org.junit.jupiter.api.Test;
 
 class FlinkStreamingRectifierTest {
 
     @Test
-    public void processDataStream() throws Exception {
-        StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
+    void processDataStream() throws Exception {
+        var env = StreamExecutionEnvironment.getExecutionEnvironment();
         env.setParallelism(2);
-        DataStream<String> rawStream = env.fromCollection(FlinkRectifierTest.source, BasicTypeInfo.STRING_TYPE_INFO);
-
-        DataStream<Row> rowStream = FlinkStreamingRectifier.operator(rawStream, new JsonSourceFormat(), FlinkRectifierTest.conf);
+        var rawStream = env.fromData(FlinkRectifierTest.source, BasicTypeInfo.STRING_TYPE_INFO);
+        var rowStream = FlinkStreamingRectifier.operator(
+                rawStream, new JsonSourceFormat(), FlinkRectifierTest.conf);
         rowStream.print();
         env.execute();
     }
